# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /shield/status:
    get:
      operationId: getShieldStatus
      responses:
        "200":
          description: Verified runtime capabilities
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RuntimeStatus" }
  /shield/settings:
    get:
      operationId: getSettings
      responses:
        "200":
          description: Session settings
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Settings" }
    put:
      operationId: saveSettings
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Settings" }
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Settings" }
  /shield/scans:
    get:
      operationId: listScans
      responses:
        "200":
          description: Session scans
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Scan" }
    post:
      operationId: startScan
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ScanInput" }
      responses:
        "202":
          description: Scan started
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Scan" }
  /shield/scans/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      operationId: getScan
      responses:
        "200":
          description: Scan and sanitized findings
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Scan" }
  /shield/scans/{id}/findings/{findingId}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
      - { name: findingId, in: path, required: true, schema: { type: string } }
    patch:
      operationId: updateFinding
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FindingUpdate" }
      responses:
        "200":
          description: Updated finding
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Finding" }
  /shield/scans/{id}/findings/{findingId}/explain:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
      - { name: findingId, in: path, required: true, schema: { type: string } }
    post:
      operationId: explainFinding
      responses:
        "200":
          description: Honest explanation provenance
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Explanation" }
  /shield/scans/{id}/report:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      operationId: getReport
      responses:
        "200":
          description: Sanitized downloadable report content
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Report" }
components:
  schemas:
    RuntimeStatus:
      type: object
      required: [gitleaksAvailable, realScanEnabled, aiAvailable, aiModel, message]
      properties:
        gitleaksAvailable: { type: boolean }
        realScanEnabled: { type: boolean }
        aiAvailable: { type: boolean }
        aiModel: { type: string }
        message: { type: string }
    Settings:
      type: object
      required: [useLocalAi, timeoutSeconds, redactEvidence, retainSource]
      properties:
        useLocalAi: { type: boolean }
        timeoutSeconds: { type: integer, minimum: 5, maximum: 120 }
        redactEvidence: { type: boolean }
        retainSource: { type: boolean }
    ScanInput:
      type: object
      required: [mode, authorized]
      properties:
        mode: { type: string, enum: [demo, local] }
        path: { type: string, maxLength: 512 }
        authorized: { type: boolean }
    FindingUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [open, resolved, ignored] }
    Finding:
      type: object
      required: [id, title, severity, category, filePath, line, evidence, status, explanation, risk, remediation, ruleId]
      properties:
        id: { type: string }
        title: { type: string }
        severity: { type: string, enum: [critical, high, medium, low] }
        category: { type: string }
        filePath: { type: string }
        line: { type: integer }
        evidence: { type: string }
        status: { type: string, enum: [open, resolved, ignored] }
        explanation: { type: string }
        risk: { type: string }
        remediation: { type: string }
        ruleId: { type: string }
    Scan:
      type: object
      required: [id, mode, repository, status, progress, stage, createdAt, durationMs, filesAnalyzed, findings, error]
      properties:
        id: { type: string }
        mode: { type: string, enum: [demo, local] }
        repository: { type: string }
        status: { type: string, enum: [running, completed, failed] }
        progress: { type: integer }
        stage: { type: string }
        createdAt: { type: string }
        durationMs: { type: integer }
        filesAnalyzed: { type: [integer, "null"] }
        findings:
          type: array
          items: { $ref: "#/components/schemas/Finding" }
        error: { type: [string, "null"] }
    Explanation:
      type: object
      required: [source, model, text, reason]
      properties:
        source: { type: string, enum: [deterministic, ollama] }
        model: { type: [string, "null"] }
        text: { type: string }
        reason: { type: string }
    Report:
      type: object
      required: [json, markdown, filename]
      properties:
        json: { type: string }
        markdown: { type: string }
        filename: { type: string }
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status