# lib/api-zod/src/generated/types/finding.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FindingSeverity } from './findingSeverity';
import type { FindingStatus } from './findingStatus';

export interface Finding {
  id: string;
  title: string;
  severity: FindingSeverity;
  category: string;
  filePath: string;
  line: number;
  evidence: string;
  status: FindingStatus;
  explanation: string;
  risk: string;
  remediation: string;
  ruleId: string;
}