# lib/api-zod/src/generated/types/runtimeStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RuntimeStatus {
  gitleaksAvailable: boolean;
  realScanEnabled: boolean;
  aiAvailable: boolean;
  aiModel: string;
  message: string;
}