# lib/api-zod/src/generated/types/scan.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Finding } from './finding';
import type { ScanMode } from './scanMode';
import type { ScanStatus } from './scanStatus';

export interface Scan {
  id: string;
  mode: ScanMode;
  repository: string;
  status: ScanStatus;
  progress: number;
  stage: string;
  createdAt: string;
  durationMs: number;
  /** @nullable */
  filesAnalyzed: number | null;
  findings: Finding[];
  /** @nullable */
  error: string | null;
}