# lib/api-zod/src/generated/types/scanInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScanInputMode } from './scanInputMode';

export interface ScanInput {
  mode: ScanInputMode;
  /** @maxLength 512 */
  path?: string;
  authorized: boolean;
}